feat: comments resolution and docs updated
