+++ internal/mapper/console_application_instance_v1/console_application_instance_v1_resource_mapper.go
@@ -7,8 +7,10 @@ import (
 	console "github.com/conduktor/terraform-provider-conduktor/internal/model/console"
 	schema "github.com/conduktor/terraform-provider-conduktor/internal/schema"
 	appinstance "github.com/conduktor/terraform-provider-conduktor/internal/schema/resource_console_application_instance_v1"
+	appinstances "github.com/conduktor/terraform-provider-conduktor/internal/schema/resource_console_application_instance_v1"
 	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
+	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
 )
 
@@ -17,7 +19,7 @@ func TFToInternalModel(ctx context.Context, r *appinstance.ConsoleApplicationIns
 	if diag.HasError() {
 		return console.ApplicationInstanceConsoleResource{}, mapper.WrapDiagError(diag, "topic_policy_ref", mapper.FromTerraform)
 	}
-	resources, err := schema.SetValueToResourceArray(ctx, r.Spec.Resources)
+	resources, err := setValueToResourceArray(ctx, r.Spec.Resources)
 	if err != nil {
 		return console.ApplicationInstanceConsoleResource{}, err
 	}
@@ -48,7 +50,7 @@ func InternalModelToTerraform(ctx context.Context, r *console.ApplicationInstanc
 		}
 	}
 
-	resourcesSet, err := schema.ResourceArrayToSetValue(ctx, r.Spec.Resources)
+	resourcesSet, err := resourceArrayToSetValue(ctx, r.Spec.Resources)
 	if err != nil {
 		return appinstance.ConsoleApplicationInstanceV1Model{}, err
 	}
@@ -80,4 +82,67 @@ func InternalModelToTerraform(ctx context.Context, r *console.ApplicationInstanc
 		Application: schema.NewStringValue(r.Metadata.Application),
 		Spec:        specValue,
 	}, nil
+}
+
+// Parse a Resources Array into a Set.
+func resourceArrayToSetValue(ctx context.Context, arr []console.ResourceWithOwnership) (basetypes.SetValue, error) {
+	var tfResources []attr.Value
+	var diag diag.Diagnostics
+
+	for _, p := range arr {
+		types := map[string]attr.Type{
+			"type":            basetypes.StringType{},
+			"name":            basetypes.StringType{},
+			"pattern_type":    basetypes.StringType{},
+			"connect_cluster": basetypes.StringType{},
+			"ownership_mode":  basetypes.StringType{},
+		}
+		values := map[string]attr.Value{
+			"type":            schema.NewStringValue(p.Type),
+			"name":            schema.NewStringValue(p.Name),
+			"pattern_type":    schema.NewStringValue(p.PatternType),
+			"connect_cluster": schema.NewStringValue(p.ConnectCluster),
+			"ownership_mode":  schema.NewStringValue(p.OwnershipMode),
+		}
+
+		permObj, diag := appinstances.NewResourcesValue(types, values)
+		if diag.HasError() {
+			return basetypes.SetValue{}, mapper.WrapDiagError(diag, "resources", mapper.FromTerraform)
+		}
+		tfResources = append(tfResources, permObj)
+
+	}
+
+	resourcesList, diag := types.SetValue(appinstances.ResourcesValue{}.Type(ctx), tfResources)
+	if diag.HasError() {
+		return basetypes.SetValue{}, mapper.WrapDiagError(diag, "resources", mapper.FromTerraform)
+	}
+
+	return resourcesList, nil
+}
+
+// Parse a Set into an array of Resources.
+func setValueToResourceArray(ctx context.Context, set basetypes.SetValue) ([]console.ResourceWithOwnership, error) {
+	resources := make([]console.ResourceWithOwnership, 0)
+	var diag diag.Diagnostics
+
+	if !set.IsNull() && !set.IsUnknown() {
+		var tfResources []appinstances.ResourcesValue
+		diag = set.ElementsAs(ctx, &tfResources, false)
+		if diag.HasError() {
+			return nil, mapper.WrapDiagError(diag, "resources", mapper.FromTerraform)
+		}
+
+		for _, p := range tfResources {
+			resources = append(resources, console.ResourceWithOwnership{
+				Type:           p.ResourcesType.ValueString(),
+				Name:           p.Name.ValueString(),
+				PatternType:    p.PatternType.ValueString(),
+				ConnectCluster: p.ConnectCluster.ValueString(),
+				OwnershipMode:  p.OwnershipMode.ValueString(),
+			})
+		}
+
+	}
+	return resources, nil
 }
